requirements.txt
pytest
pytest-mock
PyYAML
pybars3
WeasyPrint
boto3